test(differential): share the corpus sweep with test_refusal_collapsed_agreement_2454 too (#2723)
The second file the issue names spends its whole 67 s in ONE case,
`test_the_payload_gains_no_definition_marker`, which runs the identical
full sweep of the real script and asserts on its `@@` metadata rows — the
same artifact the 2345 readers now share. So `CorpusCache` and the raw
runners move out of the 2345 test module into
`python/tests/differential_corpus_2723.py` (the repo's bare-module helper
pattern, cf. `adr027_flag.py`), the `corpus` session fixture moves to
`python/tests/conftest.py`, and both modules import it. The cache's own
tests now patch the runner on the helper module, where the cache looks it up.

Verified in one session: the 2454 case pays the sweep (59 s), the 2345
sweep readers read it (0.04 s), one entry on disk; both files together
106 s wall clock vs 371 s + 61 s before. Alone, the 2454 file is unchanged
(one sweep either way). Gate-off: giving the script a `@@definition` row
turns the 2454 case red on the shared artifact while the 2345 readers,
which do not assert on metadata, stay green.
